starting typescript
