+++ client/components/Header.tsx
@@ -1,5 +1,7 @@
-import React from "react";
-function Header() {
+import React, { ReactElement } from "react";
+import { FC } from "react";
+
+const Header : FC = (): ReactElement => {
   return (
     <header className="bg-black text-white shadow-md">
       <div className="max-w-7xl mx-auto px-4 py-2 flex justify-between items-center">
