+++ package.json
@@ -38,6 +38,10 @@
     "@babel/core": "^7.24.6",
     "@babel/preset-env": "^7.24.6",
     "@babel/preset-react": "^7.24.6",
+    "@types/express": "^4.17.21",
+    "@types/node": "^20.14.4",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
     "autoprefixer": "^10.4.19",
     "babel-loader": "^9.1.3",
     "body-parser": "^1.20.2",
@@ -53,6 +57,8 @@
     "react-router-dom": "^6.22.3",
     "style-loader": "^4.0.0",
     "tailwindcss": "^3.4.4",
+    "ts-loader": "^9.5.1",
+    "typescript": "^5.4.5",
     "webpack": "^5.91.0",
     "webpack-bundle-analyzer": "^4.10.2",
     "webpack-cli": "^5.1.4",
