+++ server/output/redis-0-1.conf
@@ -1,18 +1,18 @@
 port 6380
 daemonize no
 loglevel notice
-timeout 3000
-save 0 1
+timeout 3
+save 3 1
 appendonly no
 appendfsync everysec
 rdbcompression no
 rdbchecksum no
 replica-serve-stale-data no
-maxmemory 30000
+maxmemory 5368709120
 maxmemory-policy noeviction
 # Authentication
-masterauth 
-requirepass 
+masterauth password
+requirepass password
 cluster-enabled yes
 cluster-config-file nodes.conf
 bind 0.0.0.0
