--- server/output/redis-2-1.conf
@@ -1,18 +1,0 @@
-port 6386
-daemonize no
-loglevel notice
-timeout 3000
-save 0 1
-appendonly no
-appendfsync everysec
-rdbcompression no
-rdbchecksum no
-replica-serve-stale-data no
-maxmemory 30000
-maxmemory-policy noeviction
-# Authentication
-masterauth 
-requirepass 
-cluster-enabled yes
-cluster-config-file nodes.conf
-bind 0.0.0.0
